+++ server/app.js
@@ -20,9 +20,8 @@ import WebpackHotMiddleware from 'webpack-hot-middleware';
 // Importando configurador de plantilas
 import templateEngineConfigurator from './config/templateEngine';
 
-import indexRouter from './routes/index';
-import usersRouter from './routes/users';
-import aboutRouter from './routes/about';
+// Importando enrutador principal
+import router from './routes/router';
 
 // Importando nuestro logger winston
 import winston from './config/winston';
@@ -82,9 +81,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, '..', 'public')));
 
 // Registrando las rutas en la APP
-app.use('/', indexRouter);
-app.use('/users', usersRouter);
-app.use('/about', aboutRouter);
+router.addRoutes(app);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
